목적에 따라 api를 분리한다
 - 기존 upscale, uncrop post api를 프론트에서 호출할 수 있도록 응답값을
   json으로 만들어 넘기도록 고친다
 - 기존에 테스트 목적으로 만들었던 api(프론트용으로 바꾸기 전 api
   포함해서)는 /test 하위 url로 옮겨서 계속 쓸 수 있도록 한다
