+++ src/main/java/org/rogarithm/presize/web/ImgPolishController.java
@@ -10,11 +10,10 @@ import org.rogarithm.presize.web.response.ImgUpscaleResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
+@ResponseBody
 @RequestMapping
 public class ImgPolishController {
     private static final Logger log = LoggerFactory.getLogger(ImgPolishController.class);
@@ -25,52 +24,15 @@ public class ImgPolishController {
         this.service = service;
     }
 
-    @GetMapping("/upload")
-    public String newFile(Model model, ImgUpscaleRequest upscaleRequest, ImgUncropRequest uncropRequest) {
-        model.addAttribute("ImgUpscaleRequest", upscaleRequest);
-        model.addAttribute("ImgUncropRequest", uncropRequest);
-        return "upload-img";
-    }
-
-    @GetMapping("/upscale")
-    public String showResult(@ModelAttribute("resizedImg") String resizedImg, Model model) {
-        model.addAttribute("resizedImg", resizedImg);
-        return "upscale-result";
-    }
-
-    @GetMapping("/uncrop")
-    public String showUncropResult(@ModelAttribute("uncropImg") String uncropImg, Model model) {
-        model.addAttribute("uncropImg", uncropImg);
-        return "uncrop-result";
-    }
-
     @PostMapping("/upscale")
-    public String uploadImg(@ModelAttribute("ImgUpscaleRequest") ImgUpscaleRequest request, RedirectAttributes redirectAttributes) {
+    public ImgUpscaleResponse uploadImg(@ModelAttribute("ImgUpscaleRequest") ImgUpscaleRequest request) {
         ImgUpscaleDto from = ImgUpscaleDto.from(request);
-        ImgUpscaleResponse response = service.uploadImg(from);
-
-        if (response.isSuccess()) {
-            redirectAttributes.addFlashAttribute("originalImg", from.getImg());
-            redirectAttributes.addFlashAttribute("resizedImg", response.getResizedImg());
-            return "redirect:upscale";
-        } else {
-            redirectAttributes.addFlashAttribute("error", "Image processing failed: " + response.getMessage());
-            return "redirect:upload";
-        }
+        return service.uploadImg(from);
     }
 
     @PostMapping("/uncrop")
-    public String uncropImg(@ModelAttribute("ImgUncropRequest") ImgUncropRequest request, RedirectAttributes redirectAttributes) {
+    public ImgUncropResponse uncropImg(@ModelAttribute("ImgUncropRequest") ImgUncropRequest request) {
         ImgUncropDto from = ImgUncropDto.from(request);
-        ImgUncropResponse response = service.uncropImg(from);
-
-        if (response.isSuccess()) {
-            redirectAttributes.addFlashAttribute("originalImg", from.getImg());
-            redirectAttributes.addFlashAttribute("uncropImg", response.getUncropImg());
-            return "redirect:uncrop";
-        } else {
-            redirectAttributes.addFlashAttribute("error", "Image processing failed: " + response.getMessage());
-            return "redirect:upload";
-        }
+        return service.uncropImg(from);
     }
 }
