+++ src/main/resources/templates/uncrop-result.html
@@ -17,7 +17,7 @@
             <img th:src="'data:image;base64,' + ${uncropImg}" alt="변환된 이미지"/>
         </div>
     </div>
-    <a href="/upload">다른 이미지 변환하기</a>
+    <a href="/test/upload">다른 이미지 변환하기</a>
 </div>
 </body>
 </html>
